% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Función 2.R
\name{calcular_costo_variable_unitario}
\alias{calcular_costo_variable_unitario}
\title{Función para calcular el costo variable por unidad}
\usage{
calcular_costo_variable_unitario(costos_variables, unidades_producidas)
}
\arguments{
\item{costos_variables}{(varían en proporción a la cantidad de unidades producidas)}

\item{unidades_producidas}{(cantidad de unidades)}
}
\value{
El costo por unidad producida; el valor del costo variable unitario calculado.
}
\description{
Función para calcular el costo variable por unidad
}
