% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Función 1.R
\name{calcular_flujo_caja}
\alias{calcular_flujo_caja}
\title{Calculadora de para la producción de carne de conejos.}
\usage{
calcular_flujo_caja(costos_fijos, costos_variables, ingresos)
}
\arguments{
\item{costos_fijos}{(representa los costos fijos asociados a la producción. Estos costos no varían con la cantidad de unidades producidas)}

\item{costos_variables}{(epresenta los costos variables asociados a la producción. Estos costos dependen de la cantidad de unidades producidas y varían en proporción a dicha cantidad)}

\item{ingresos}{(ingresos generados por la venta de unidades producidas)}
}
\value{
El valor calculado como resultado de la función.
}
\description{
Función para calcular el flujo de caja.
}
