% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Función 3.R
\name{calcular_punto_equilibrio}
\alias{calcular_punto_equilibrio}
\title{Función para calcular el punto de equilibrio}
\usage{
calcular_punto_equilibrio(costos_fijos, costo_variable_unitario)
}
\arguments{
\item{costos_fijos}{(Estos costos no varían con la cantidad de unidades producidas)}

\item{costo_variable_unitario}{(el costo promedio por unidad producida)}
}
\value{
El valor del punto de equilibrio calculado, que representa la cantidad de unidades a vender para alcanzar el equilibrio financiero
}
\description{
Función para calcular el punto de equilibrio
}
