% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funci贸n 4.R
\name{calcular_roi}
\alias{calcular_roi}
\title{Funci贸n para calcular el ROI}
\usage{
calcular_roi(ingresos, costos_totales)
}
\arguments{
\item{ingresos}{(ingresos generados por la venta de las unidades producidas)}

\item{costos_totales}{(costos totales, que incluyen los costos fijos y los costos variables)}
}
\value{
valor del ROI calculado, que representa el porcentaje de retorno sobre la inversi贸n realizada.
}
\description{
Funci贸n para calcular el ROI
}
