% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Función 5.R
\name{calcular_roi2}
\alias{calcular_roi2}
\title{Función para calcular el ROI2 para un escenario específico}
\usage{
calcular_roi2(demanda, costos_insumos)
}
\arguments{
\item{demanda}{(demanda del producto o servicio. Puede ser un valor numérico, una variable o una fórmula que indique la demanda esperada)}

\item{costos_insumos}{(costos de insumos o materiales necesarios para la producción del producto o servicio)}
}
\description{
Función para calcular el ROI2 para un escenario específico
}
