% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Función 6.R
\name{realizar_analisis_escenarios}
\alias{realizar_analisis_escenarios}
\title{Función para realizar el análisis de escenarios}
\usage{
realizar_analisis_escenarios(escenarios)
}
\description{
Función para realizar el análisis de escenarios
}
